Issue 26 — Shared API Client, Auth Guard, and Route Protection Fixes #32
