+++ frontend/src/components/Login.jsx
@@ -1,0 +1,49 @@
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAuth } from '../contexts/AuthContext';
+
+const Login = () => {
+  const [token, setToken] = useState('');
+  const [role, setRole] = useState('student');
+  const { login } = useAuth();
+  const navigate = useNavigate();
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (!token.trim()) return;
+    // For now, simulate user data based on role
+    const userData = { role, id: '123' }; // TODO: Fetch from backend
+    login(token, userData);
+    navigate('/dashboard');
+  };
+
+  return (
+    <div className="login-page">
+      <h1>Login</h1>
+      <form onSubmit={handleSubmit}>
+        <label>
+          Token:
+          <input
+            type="text"
+            value={token}
+            onChange={(e) => setToken(e.target.value)}
+            placeholder="Enter JWT token"
+            required
+          />
+        </label>
+        <label>
+          Role:
+          <select value={role} onChange={(e) => setRole(e.target.value)}>
+            <option value="student">Student</option>
+            <option value="coordinator">Coordinator</option>
+            <option value="admin">Admin</option>
+            <option value="professor">Professor</option>
+          </select>
+        </label>
+        <button type="submit">Login</button>
+      </form>
+    </div>
+  );
+};
+
+export default Login;
