+++ frontend/src/contexts/AuthContext.jsx
@@ -1,0 +1,59 @@
+import { createContext, useContext, useState, useEffect } from 'react';
+
+const AuthContext = createContext();
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
+
+export const AuthProvider = ({ children }) => {
+  const [token, setToken] = useState(null);
+  const [user, setUser] = useState(null); // { role, id, etc. }
+
+  // Load token from localStorage on mount
+  useEffect(() => {
+    const storedToken = localStorage.getItem('authToken');
+    if (storedToken) {
+      setToken(storedToken);
+      // TODO: Optionally fetch user info from backend using token
+    }
+  }, []);
+
+  const login = (newToken, userData) => {
+    setToken(newToken);
+    setUser(userData);
+    localStorage.setItem('authToken', newToken);
+  };
+
+  const logout = () => {
+    setToken(null);
+    setUser(null);
+    localStorage.removeItem('authToken');
+  };
+
+  const isAuthenticated = !!token;
+
+  const hasRole = (requiredRoles) => {
+    if (!user || !user.role) return false;
+    return requiredRoles.includes(user.role);
+  };
+
+  const value = {
+    token,
+    user,
+    isAuthenticated,
+    hasRole,
+    login,
+    logout,
+  };
+
+  return (
+    <AuthContext.Provider value={value}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
